chore: move non grunt scripts to scripts
