+++ .npmignore
@@ -3,6 +3,7 @@
 tmp
 test
 tasks
+scripts
 docs
 
 TODO.md
