--- init-dev-env.bat
@@ -1,1 +1,0 @@
-node init-dev-env.js
