+++ scripts/init-dev-env.bat
@@ -1,0 +1,2 @@
+set __dirname=%~dp0
+node %__dirname%init-dev-env.js
